requirements.txt
flask>=2.0.1
google-generativeai>=0.3.0
python-dotenv>=0.19.0
flask-sqlalchemy>=3.0.0
sqlalchemy>=2.0.0
werkzeug>=2.0.0
flask-login>=0.6.0
flask-migrate>=4.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
pygments>=2.15.0
black>=23.0.0
flask-socketio==5.3.6
shortuuid==1.0.11
eventlet==0.33.3
flask-sse==1.0.0
redis==5.0.1
